[WIP] Starting transformation process
The grammar includes expressions that allow values to be captured during validation. If all properties in a propertyGroup match, then those values are used in the constructor assigned to the propertyDescription (need better terms here).

Added ability for user to supply functions to validate types unknown to the validation system and to create new instances. These functions can be handled by specific type names or a single general purpose handler for each category can be registered using ‘*’ the type name.