Lots of parser changes
This is starting to differentiate between top-level type descriptions and type patterns used for matching inside of a ‘match’ block. As a result, the previous array support has been moved down to the type pattern level. We only support empty arrays at the top level for now.

This adds support for enumeration. You can make a list of string constants and test a string’s value against that list. If the value is in the list, the match succeeds. Right now, this only returns the matched value.

Add support for matching any kind of array and any kind of object.

Strings may be used for property names when those names match a keyword. This also allows for names with whitespace in them.

We now require groups to have at least one match. I think this makes sense semantically and it helps the logic in the match handling code.

We now differentiate between integers and floats. For example, ranges require integers; whereas, a type can be defined as matching a specific float value.