Add support for null/undefined
These work as top-level types and as type patterns. Note that the library currently uses undefined to report an error. There is a special sigil within Normalizer that can be changed to use a unique object instead; however, I will need to make sure that doesn’t break return values…holding off for now as this is most likely a low-priority, little-used type