Remove type and use spread args
This allows for more natrual-looking functions used as type creators. Note that this means it will be less likely that a single function can be used for multiple type handlers, but I think that’s OK. You can still use “*” as a general handler. That function receives the type and args as a separate argument as specific handles did before.