Start CLI tool
This allows you to specify a norm file to load, optionally a module to load (to provide type creators), a JSON file to load and the type to use to process the JSON data. If the JSON data is an array, each item in the array will be processed separately.