Allow patterns and transforms to be stored by name
This allows us to reuse patterns ands transforms by using the name in the assignment in place of the pattern/transform.

I’m a little on the fence on this. It is obviously very useful for DRY, but I’m a little concerned that it separates the context between where captures are defined and where they are used. Hopefully, these transformation scripts won’t get too large (famous last words?)