Fix element group capture logic
This special cases when we have a range from 1..1. We prevent creating the extra layer of arrays, which feels more natural. However, that exrra level is absolutely required when we match more than on group. We assume that if the repetition ends any higher than 1 that we will get more than one capture even though we may not.