Completely rework grammar
There is too much to capture here, but the basic idea is that we can now flow expressions (previously called generators) and patterns into one another. We can define variables at the top-level scope, in an array scope, and in object scope.

Still todo is to get the kld-intersections transformations working again. I’ll need to add a few more constructs to get that working, but it’s pretty close.